great news! -- selected/cell fixing
So I accidentally used `selected` where `cell` should have been. The good news is, the Cell ID is translated to the back-end, and I have a way to index by ID. That means that getting control back to the cells with this new keymap technique (where we don't bind each cell individually) should actually be relatively smooth and painless.
Other than this, there are a few things -- syntax highlighting, general polish, TABS (yes, this is coming), finishing the explorer (and allowing it to open files into a current session), and finally the settings panel. Then it is just docs and it is ready for release -- that is once these cell controls are fully realized. Also notably need to get the drag handle working... And generally make the environment a lot more draggable.