more cell updates; keybinds brought into cells
the `bind!` method from toolips session does not necessarily work for this.... Yet. However, soon I will figure this all out and then finally these cells can be " put together" for the **last time**. I don't know if you have realized this yet from my commit messages, but oh my goodness gracious this project has become complicated and a pain to work on. The pay-off is huge though -- so don't think I am giving up or anything, it is just going to be an absurd amount of work to get all of this done. Eventually it certainly will be!