finished main setup screen
so this is the main home selector screen, which lets you select your home directory for olive. I still need to setup some sort of source location configuration that keeps this location in a predictable place. Also, need to make a " do you want to add olive defaults?" and the olive tutorial portion.