added TextModifier highlighter, doc route (**read**)
still working on the `TextModifier`, but this will be easy and toolips-effective thanks to Components. The docbrowser route is a new full-blown documentation browser. Eventually, I plan to have a full blown module explorer here. For now, the focus is fixing the indentation and gussying up the cells... Things like that. I will be registering `0.0.8` with only a single-use doc browser for now. I need a function to build different documentation. I might even connect it to a Symbol, lol