fixed `Directory` cells sometimes not existing
Prior, directory functions would assume that a path is a file if it is not a directory. The problem is when we have a path which is not a file or a directory. For this case, I will give up on providing that file or directory and move onto the rest of the files or directories.