extracted main, made keys options, more `readonly`
This will improve on so many different things.
- revisions to `open_project!`
- revisions to main, allowing for the option of turning off key verification. Also simplified the route with more methods
- simplified the stylesheet, moved some styles around

Next, I will be cutting down the code of the cells -- mainly the REPL cells, I am also going to drastically improve the markdown cells, the process of activating a project, and viewing its environment. Etc. I have a lot, but this closes.... Lemme look this up.
#108
#91
#90
At least from a base-level perspective. The ability to use `drag` to change the pane, for example in #91 is not there yet, but