another substantial update into `0.3`
Fixed up the abstraction of some things, resulting in a lot more things working. Got extensions working, and updated the version output on startup. I am going to add real regular buttons to the `getstarted` cell for the issues and yadda yadda. Things are looking good, I like this :)) My next mission will be to move the stylesheet to an editable system on a client-client basis. this will be alongside trying to work through any sorts of bugs, and wrapping up some updates to different cells. A new version of `ToolipsServables`, `Toolips`, and `ParametricProcesses` are also coming ;)