more `:module` cell updates
These are proving to be challenging; quite a difficult concept to accomplish, but this code is getting really close. With this, essentially every `Module` cell will be working really well. I also plan to do some more work to `get_started`. From there, it is all about fixing any bugs that might remain and increasing efficiency. The biggest thing will be changing from inline styles to a more stylesheet-based approach, including a settings menu for editing the styles. Cutting down on the time it takes to build a cell both in that way and other efficiency ways, for example many of the indexes will call `getfield` over and over again when it makes more sense to allocate. All in all, this is getting really close to being *the build* -- BETA -- the step that will take us into the era of `OliveCreator` and more. Exciting!