added new `focus_on!` and `focus_off!` system.
This makes it easier to support **convenient** focus for a wider variety of cells -- we need a parametric function here because we want the functionality to change in certain cases. A great example is `:markdown` where the cell will need to become editable and ineditable as we shift focus across it. This will also later be implemented for `cell_up!` and `cell_down!`