THEMES MVP!
So -- THEMES are now working ! (completely). Upon activating themes from the settings (for the first time,) a themes directory will be built within your `olive` home. This theme directory will contain `.olivestyle` files, which have a custom binding to open within `Olive` these styles open as `tomlvalues` cells and the save automatically composes them into an instantly accessible `String`. The whole implementation is nice, and pretty lax on memory compared to the alternative. I really like the way it came out! The one significant shortcoming to the current iteration is that it only works with the `Style` type directly. This is fine for the most part, but will exclude `KeyFrames` and as a result make the page far less animated.